requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
accelerate=1.11.0=pypi_0
acres=0.5.0=pypi_0
aiofiles=25.1.0=pypi_0
altair=5.5.0=pypi_0
anyio=4.10.0=py310haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=25.1.0=py310h02ab6af_0
asttokens=3.0.0=py310haa95532_0
async-lru=2.0.5=py310haa95532_0
attrs=25.4.0=py310haa95532_2
babel=2.16.0=py310haa95532_0
beautifulsoup4=4.13.5=py310haa95532_0
bleach=6.2.0=py310haa95532_0
blinker=1.9.0=pypi_0
brotlicffi=1.0.9.2=py310h885b0b7_2
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.11.4=haa95532_0
cachetools=6.2.1=pypi_0
certifi=2025.10.5=py310haa95532_0
cffi=2.0.0=py310h02ab6af_1
charset-normalizer=3.4.4=py310haa95532_0
ci-info=0.3.0=pypi_0
click=8.3.0=pypi_0
colorama=0.4.6=py310haa95532_0
comm=0.2.1=py310haa95532_0
configobj=5.0.9=pypi_0
configparser=7.2.0=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.16=py310h885b0b7_1
decorator=5.2.1=py310haa95532_0
defusedxml=0.7.1=pyhd3eb1b0_0
etelemetry=0.3.1=pypi_0
exceptiongroup=1.3.0=py310haa95532_0
executing=2.2.1=py310haa95532_0
expat=2.7.3=h9214b88_0
filelock=3.19.1=pypi_0
fonttools=4.60.1=pypi_0
frontend=0.0.3=pypi_0
fsspec=2025.9.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
h11=0.16.0=py310haa95532_1
httpcore=1.0.9=py310haa95532_0
httplib2=0.31.0=pypi_0
httpx=0.28.1=py310haa95532_1
huggingface-hub=0.36.0=pypi_0
idna=3.11=py310haa95532_0
importlib-resources=6.5.2=pypi_0
ipykernel=6.30.1=py310haa95532_0
ipython=8.30.0=py310haa95532_0
isodate=0.7.2=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=py310haa95532_0
jinja2=3.1.6=py310haa95532_0
joblib=1.5.2=pypi_0
json5=0.9.25=py310haa95532_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
jupyter-lsp=2.2.5=py310haa95532_0
jupyter_client=8.6.3=py310haa95532_0
jupyter_core=5.8.1=py310haa95532_0
jupyter_events=0.12.0=py310haa95532_0
jupyter_server=2.16.0=py310haa95532_0
jupyter_server_terminals=0.5.3=py310haa95532_0
jupyterlab=4.4.7=py310haa95532_0
jupyterlab_pygments=0.3.0=py310haa95532_0
jupyterlab_server=2.28.0=py310haa95532_0
kiwisolver=1.4.9=pypi_0
libffi=3.4.4=hd77b12b_1
libsodium=1.0.20=h83e8143_0
libzlib=1.3.1=h02ab6af_0
looseversion=1.3.0=pypi_0
lxml=6.0.2=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.7=pypi_0
matplotlib-inline=0.1.7=py310haa95532_0
mistune=3.1.2=py310haa95532_0
mpmath=1.3.0=pypi_0
narwhals=2.10.2=pypi_0
nbclient=0.10.2=py310haa95532_0
nbconvert-core=7.16.6=py310haa95532_0
nbformat=5.10.4=py310haa95532_0
nest-asyncio=1.6.0=py310haa95532_0
networkx=3.3=pypi_0
nibabel=5.3.2=pypi_0
nipype=1.10.0=pypi_0
notebook-shim=0.2.4=py310haa95532_0
numpy=2.1.2=pypi_0
opencv-python=4.12.0.88=pypi_0
openssl=3.0.18=h543e019_0
overrides=7.7.0=py310haa95532_0
packaging=25.0=py310haa95532_1
pandas=2.3.3=pypi_0
pandocfilters=1.5.1=py310haa95532_0
parso=0.8.4=py310haa95532_0
pathlib=1.0.1=pypi_0
pillow=11.3.0=pypi_0
pip=25.2=pyhc872135_1
platformdirs=4.5.0=py310haa95532_0
prometheus_client=0.21.1=py310haa95532_0
prompt-toolkit=3.0.52=py310haa95532_1
prompt_toolkit=3.0.52=hd3eb1b0_1
protobuf=6.33.0=pypi_0
prov=2.1.1=pypi_0
psutil=7.1.3=pypi_0
pure_eval=0.2.3=py310haa95532_0
puremagic=1.30=pypi_0
pyarrow=21.0.0=pypi_0
pycparser=2.23=py310haa95532_0
pydeck=0.9.1=pypi_0
pydot=4.0.1=pypi_0
pygments=2.19.2=py310haa95532_0
pymupdf=1.26.6=pypi_0
pyparsing=3.2.5=pypi_0
pysocks=1.7.1=py310haa95532_1
pytesseract=0.3.13=pypi_0
python=3.10.19=h981015d_0
python-dateutil=2.9.0post0=py310haa95532_2
python-fastjsonschema=2.20.0=py310haa95532_0
python-json-logger=3.2.1=py310haa95532_0
pytz=2025.2=pypi_0
pywin32=311=py310h885b0b7_0
pywinpty=2.0.15=py310h72d21ff_0
pyxnat=1.6.4=pypi_0
pyyaml=6.0.3=pypi_0
pyzmq=27.1.0=py310h7149c55_1
rdflib=7.4.0=pypi_0
referencing=0.37.0=py310haa95532_0
regex=2025.11.3=pypi_0
requests=2.32.5=py310haa95532_1
rfc3339-validator=0.1.4=py310haa95532_0
rfc3986-validator=0.1.1=py310haa95532_0
rpds-py=0.28.0=py310h114bc41_0
safetensors=0.6.2=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.2=py310haa95532_1
setuptools=80.9.0=py310haa95532_0
simplejson=3.20.2=pypi_0
six=1.17.0=py310haa95532_0
smmap=5.0.2=pypi_0
sniffio=1.3.0=py310haa95532_0
soupsieve=2.5=py310haa95532_0
sqlite=3.51.0=hda9a48d_0
stack_data=0.6.3=py310haa95532_0
starlette=0.50.0=pypi_0
streamlit=1.51.0=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.1.2=pypi_0
terminado=0.18.1=py310haa95532_0
threadpoolctl=3.6.0=pypi_0
tinycss2=1.4.0=py310haa95532_0
tk=8.6.15=hf199647_0
tokenizers=0.19.1=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=py310haa95532_0
tools=1.0.10=pypi_0
torch=2.6.0+cu118=pypi_0
torchaudio=2.6.0+cu118=pypi_0
torchvision=0.21.0+cu118=pypi_0
tornado=6.5.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=py310haa95532_0
traits=7.0.2=pypi_0
transformers=4.41.0=pypi_0
typing-extensions=4.15.0=py310haa95532_0
typing_extensions=4.15.0=py310haa95532_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.5.0=py310haa95532_0
uvicorn=0.38.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=py310haa95532_0
webencodings=0.5.1=py310haa95532_1
websocket-client=1.8.0=py310haa95532_0
wheel=0.45.1=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_1
winpty=0.4.3=4
xz=5.6.4=h4754444_1
yaml=0.2.5=he774522_0
zeromq=4.3.5=h6c54ac7_1
zlib=1.3.1=h02ab6af_0
